Move error to new decode module
Observe that the `consensus::encode::Error` is actually an error
encountered while decoding.

Create a new submodule `consensus::decode` and move the error there,
re-export the error from `encode` still so this is not a breaking
change.
